remove gitignore. Move creation address of loc-config to ~/. Update prepare to test whether config exists beforehand. Update install to move scripts to usr/local/bin.

+++ loc-prepare.sh
@@ -1,12 +1,17 @@
 #!/bin/bash
 #
-# The prepare script simply creates a copy of the desired oral history folder from Seed_Bank parent directory to LOC_PreRelease directory.
+# The prepare script simply creates a copy of the desired oral history folder from Seed_Bank parent directory to LOC_PreRelease directory. Prepends oral history folder name with temp_.
 
-source loc-config
-target=$LOC_PreRelease
+if [[ -f ~/loc-config ]]; then
+  source ~/loc-config
+  target=$LOC_PreRelease
 
-for i in "$@"
-do
-  echo Copying ${i} to ${target}
-  cp -R ${i} "$target/temp_$i"
-done
+  for i in "$@"
+  do
+    echo Copying ${i} to ${target}
+    cp -R ${i} "$target/temp_$i"
+  done
+else
+  echo "Couldn't find loc-config. Please run loc-setup."
+fi
+
